requirements.txt
streamlit>=1.38.0
pandas>=2.1.4
pillow>=10.3.0
matplotlib>=3.8.0
seaborn>=0.13.2
